main/requirements.txt
requests
python-dotenv
langchain
langchain-openai
langchain-core
langchain-community
pydantic
duckduckgo-search